read trip with combined conditions

+++ trips/src/main/java/be/vinci/ipl/chattycar/trips/models/Position.java
@@ -12,6 +12,6 @@ import lombok.ToString;
 @NoArgsConstructor
 @Embeddable
 public class Position {
-  private float latitude;
-  private float longitude;
+  private Double latitude;
+  private Double longitude;
 }
